Add basic AppStream metainfo data
For packaging the Red Hat fonts in Linux distributions (such as Fedora),
AppStream metadata is required to support discoverability and installability
through modern software centers like GNOME Software and Plasma Discover.

This adds the most basic data for supporting that, which can be fleshed
out more over time as needed.